use AWS' aws-nitro-enclaves-image-format instead of maintaining our own CLI fork (#24)
Thanks to https://github.com/aws/aws-nitro-enclaves-image-format/pull/27
, we can now build EIFs via a low-level CLI that uses
`aws-nitro-enclave-image-format` directly, rather than maintaining and
developing our own CLI (which was forked from an old example in that
repo anyway).

For us, this means we can remove all the Rust code in this repository
and defer to AWS.


So this PR:
- packages AWS' `eif_build` with Nix to use as build dependency
- replaces `eif-cli` with `eif_build`, removing `eif-cli` altogether
- fixes outdated code comment and flake description
- adds a Cargo.lock to lock upstream's dependencies. This is needed to
reproducibly build `eif_build`. We can remove it in the future if
https://github.com/aws/aws-nitro-enclaves-image-format/issues/34 is
closed.

**Note EIFs built after this PR are not bit-by-bit identical to the ones
built before this PR** (so they will not have the same SHA256 digests).
This is because the EIF metadata has been updated too.

PCRs, on the other hand, can be expected to remain the same (because
metadata is not included in PCR0 hashing). We assert this via our [flake
check](https://github.com/monzo/aws-nitro-util/actions/runs/10558127694/job/29246984165?pr=24)
, which builds a test EIF against a hardcoded PCR0.

As far as downstream users are concerned, this PR has no other breaking
changes.